+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - Follow directory standards. (#79)
 - Add --index-key which will select an index based on its entry in the config file. (#78)
 - Add --config to allow selection of a config file anywhere. (#80)
+- Use panics instead of using std::process::exit. (#47)
 
 ## [0.5.3](https://github.com/cargo-prebuilt/cargo-prebuilt/releases/tag/v0.5.3)